Ensure that no collisions occur
Previously, we had many collisions and even very early:

    new: [ "3DAZ" ] = 1157 old: [ "3DAZ" ] = 119
    new: [ "5rNG" ] = 1309 old: [ "5rNG" ] = 1254
    new: [ "39cY" ] = 1916 old: [ "39cY" ] = 1509
    new: [ "SpB" ] = 2137 old: [ "SpB" ] = 1146
    new: [ "5UUN" ] = 2157 old: [ "5UUN" ] = 1827
    new: [ "2ipq" ] = 2202 old: [ "2ipq" ] = 1961
    new: [ "22PU" ] = 3202 old: [ "22PU" ] = 2874
    new: [ "5BEj" ] = 3494 old: [ "5BEj" ] = 297
    new: [ "4kUW" ] = 3718 old: [ "4kUW" ] = 2508

Now I am using a complete 32-bit value, the ids are slightly longer
because of this.

+++ Models/DatabaseModel.hpp
@@ -79,6 +79,6 @@ namespace Accounting::Models
         BillListModel *m_billList_model;
         PocketMoneyProxyModel *m_pocketMoney_model;
 
-        quint64 m_next_id = 1;
+        quint32 m_next_id = 1;
     };
 }
